% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{associate}
\alias{associate}
\title{Statistical test of association between arbitrary vector types.}
\usage{
associate(a, b, ...)
}
\arguments{
\item{a, b}{Vectors of covariates of arbitrary type.
See Details for the specific tests used.}

\item{...}{Passed to specific methods.}
}
\value{
A p-value of an association test.
}
\description{
Statistical test of association between arbitrary vector types.
}
\details{
For numeric vs numeric, a correlation test.
For factor vs factor, a chisq test.
For numeric vs factor/character/logical, ANOVA.
}
