% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association-plot.R
\name{association_plot}
\alias{association_plot}
\alias{association_table}
\title{Graphical and tabular summaries of association between one or two tables.}
\usage{
association_plot(
  a,
  b = a,
  n = 30,
  progress_bar = FALSE,
  verbose = FALSE,
  max_na = 0.9,
  ...
)

association_table(a, b = a, progress_bar = FALSE)
}
\arguments{
\item{a}{A data.frame of covariates.}

\item{b}{An optional data.frame of covariates.}

\item{n}{An integer controlling the maximum number of associations shown.}

\item{progress_bar}{Show a progress bar when calculating associations?}

\item{verbose}{Logical flag that controls whether the indices being tested
are printed at each iteration. Useful mainly for debugging.}

\item{max_na}{The maximum proportion of missing values in a row/column before
it is removed from the plot.}

\item{...}{Passed to internal functions.}
}
\value{
A ggplot or table showing the p-values of association between table
columns.
}
\description{
Graphical and tabular summaries of association between one or two tables.
}
\details{
For information on the tests used, see \code{\link{associate}}.
}
\examples{
mat <- matrix(rnorm(1000), ncol = 10)
association_plot(mat)
## only the top 5 associations
association_plot(mat, n = 5)
}
