% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-association-plot.R
\name{pca_association_plot}
\alias{pca_association_plot}
\title{Graphical test between principal components and a table of covariates.}
\usage{
pca_association_plot(a, b, ...)
}
\arguments{
\item{a}{A matrix or data.frame-alike of covariates.}

\item{b}{A matrix, or the output of \code{\link[stats]{prcomp}}.}

\item{...}{Passed to specific methods.}
}
\value{
The output of plot_grid.
}
\description{
Graphical test between principal components and a table of covariates.
}
