% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{associate}
\alias{associate}
\alias{associate,numeric,numeric-method}
\alias{associate,factor,numeric-method}
\alias{associate,numeric,factor-method}
\alias{associate,factor,factor-method}
\alias{associate,character,ANY-method}
\alias{associate,Date,ANY-method}
\alias{associate,ANY,Date-method}
\alias{associate,ANY,POSIXct-method}
\alias{associate,POSIXct,ANY-method}
\alias{associate,ANY,POSIXt-method}
\alias{associate,POSIXt,ANY-method}
\alias{associate,character,character-method}
\alias{associate,ANY,character-method}
\alias{associate,ANY,factor-method}
\alias{associate,logical,ANY-method}
\alias{associate,ANY,logical-method}
\title{Statistical test of association between arbitrary vector types.}
\usage{
associate(a, b, ...)

\S4method{associate}{numeric,numeric}(a, b, method = "spearman")

\S4method{associate}{factor,numeric}(a, b)

\S4method{associate}{numeric,factor}(a, b)

\S4method{associate}{factor,factor}(a, b)

\S4method{associate}{character,ANY}(a, b)

\S4method{associate}{Date,ANY}(a, b)

\S4method{associate}{ANY,Date}(a, b)

\S4method{associate}{ANY,POSIXct}(a, b)

\S4method{associate}{POSIXct,ANY}(a, b)

\S4method{associate}{ANY,POSIXt}(a, b)

\S4method{associate}{POSIXt,ANY}(a, b)

\S4method{associate}{character,character}(a, b)

\S4method{associate}{ANY,character}(a, b)

\S4method{associate}{ANY,factor}(a, b)

\S4method{associate}{logical,ANY}(a, b)

\S4method{associate}{ANY,logical}(a, b)
}
\arguments{
\item{a, b}{Vectors of covariates of arbitrary type.
See Details for the specific tests used.}

\item{...}{Passed to specific methods.}

\item{method}{For numeric vs numeric, the correlation method (passed to
\code{\link[stats]{cor.test}})}
}
\value{
A p-value of an association test.
}
\description{
Statistical test of association between arbitrary vector types.
}
\details{
For numeric vs numeric, a correlation test.
For factor vs factor, a chisq test.
For numeric vs factor/character/logical, ANOVA.
}
\examples{
a <- rnorm(100)
b <- runif(100)
c <- sample(letters[1:2], 100, replace = TRUE)
d <- sample(letters[3:5], 100, replace = TRUE)
associate(a, b)
associate(b, a)
associate(b, c)
associate(b, d)
associate(c, d)
}
