% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association-plot.R
\name{association_plot}
\alias{association_plot}
\alias{association_table}
\title{Graphical and tabular summaries of association between one or two tables.}
\usage{
association_plot(a, b, progress_bar = FALSE, ...)

association_table(a, b, progress_bar = FALSE)
}
\arguments{
\item{a}{A data.frame of covariates.}

\item{b}{An optional data.frame of covariates.}

\item{progress_bar}{Show a progress bar when calculating associations?}

\item{...}{Passed to \code{pvalue_heatmap}.}
}
\value{
A ggplot or table showing the p-values of association between table
columns.
}
\description{
Graphical and tabular summaries of association between one or two tables.
}
\details{
For information on the tests used, see \code{\link{associate}}.
}
