% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximise_trace.R
\name{maximise_trace}
\alias{maximise_trace}
\title{Maximise the trace of a matrix.}
\usage{
maximise_trace(mat)
}
\arguments{
\item{mat}{Input matrix}
}
\value{
A matrix with rows permuted such that tr(mat) is maximised.
}
\description{
Maximising the trace of a matrix amounts to maximising the diagonal elements.
This is useful when comparing clustering methods, or comparing clustering 
results with known ground truth labels.
}
