% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-association-plot.R
\name{pca_association_plot}
\alias{pca_association_plot}
\alias{pca_association_plot,data.frame,matrix-method}
\alias{pca_association_plot,ANY,ANY-method}
\alias{pca_association_plot,data.frame,missing-method}
\alias{pca_association_plot,ANY,data.frame-method}
\alias{pca_association_plot,data.frame,irlba_prcomp-method}
\alias{pca_association_plot,data.frame,prcomp-method}
\title{Graphical test between principal components and a table of covariates.}
\usage{
pca_association_plot(a, b, ...)

\S4method{pca_association_plot}{data.frame,matrix}(
  a,
  b,
  method = c("irlba", "prcomp"),
  npcs = min(ncol(a) - 1, nrow(a) - 1, 50),
  center = TRUE,
  scale = TRUE,
  max_iterations = 1e+05,
  ...
)

\S4method{pca_association_plot}{ANY,ANY}(a, b, ...)

\S4method{pca_association_plot}{data.frame,missing}(a, b, ...)

\S4method{pca_association_plot}{ANY,data.frame}(a, b, ...)

\S4method{pca_association_plot}{data.frame,irlba_prcomp}(a, b, progress_bar = TRUE, npcs = ncol(b$x), ...)

\S4method{pca_association_plot}{data.frame,prcomp}(a, b, npcs = ncol(b$x), progress_bar = TRUE, ...)
}
\arguments{
\item{a}{A matrix or data.frame-alike of covariates.}

\item{b}{A matrix, or the output of \code{\link[stats]{prcomp}}.}

\item{...}{Passed to specific methods.}

\item{method}{The method used to calculate PCs. Can be \code{"irlba"}
for \code{\link[irlba]{prcomp_irlba}} for truncated PCA or
\code{\link[stats]{prcomp}} for standard PCA.}

\item{npcs}{The number of PCs to truncate the results to. For large datasets
visualising >50 PCs is unwieldy so setting this to (e.g.) 20 can be very
useful.}

\item{center, scale}{Passed to PCA methods. Should the data be scaled and
centered before performing PCA?}

\item{max_iterations}{Passed to \code{\link[irlba]{prcomp_irlba}}.}

\item{progress_bar}{Show a progress bar when testing associations? Useful
for very large datasets.}
}
\value{
The output of plot_grid.
}
\description{
Graphical test between principal components and a table of covariates.
}
