% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-association-plot.R
\name{pca_association_plot}
\alias{pca_association_plot}
\alias{pca_association_plot,data.frame,matrix-method}
\alias{pca_association_plot,ANY,ANY-method}
\alias{pca_association_plot,data.frame,missing-method}
\alias{pca_association_plot,ANY,data.frame-method}
\alias{pca_association_plot,data.frame,irlba_prcomp-method}
\alias{pca_association_plot,data.frame,prcomp-method}
\title{Graphical test between principal components and a table of covariates.}
\usage{
pca_association_plot(a, b, ...)

\S4method{pca_association_plot}{data.frame,matrix}(
  a,
  b,
  method = c("irlba", "prcomp"),
  n = 20,
  npcs = min(ncol(a) - 1, nrow(a) - 1, 50),
  center = TRUE,
  scale = TRUE,
  max_iterations = 1e+05,
  vars_keep = NULL,
  vars_ignore = NULL,
  ...
)

\S4method{pca_association_plot}{ANY,ANY}(a, b, ...)

\S4method{pca_association_plot}{data.frame,missing}(a, b, ...)

\S4method{pca_association_plot}{ANY,data.frame}(a, b, ...)

\S4method{pca_association_plot}{data.frame,irlba_prcomp}(
  a,
  b,
  npcs = ncol(b$x),
  ncovariates = 20,
  vars_keep = NULL,
  vars_ignore = NULL,
  progress_bar = FALSE,
  ...
)

\S4method{pca_association_plot}{data.frame,prcomp}(
  a,
  b,
  npcs = ncol(b$x),
  ncovariates = 20,
  vars_keep = NULL,
  vars_ignore = NULL,
  progress_bar = FALSE,
  ...
)
}
\arguments{
\item{a}{A matrix or data.frame-alike of covariates.}

\item{b}{A matrix, or the output of \code{\link[stats]{prcomp}}.}

\item{...}{Passed to specific methods.}

\item{method}{The method used to calculate PCs. Can be \code{"irlba"}
for \code{\link[irlba]{prcomp_irlba}} for truncated PCA or
\code{\link[stats]{prcomp}} for standard PCA.}

\item{n}{The maximum number of variables to be shown in the PCA plot
(the top n are taken based on the minimum association p-value).}

\item{npcs}{The number of PCs to truncate the results to. For large datasets
visualising >50 PCs is unwieldy so setting this to (e.g.) 20 can be very
useful.}

\item{center, scale}{Passed to PCA methods. Should the data be scaled and
centered before performing PCA?}

\item{max_iterations}{Passed to \code{\link[irlba]{prcomp_irlba}}.}

\item{vars_keep, vars_ignore}{Vectors of variables to keep or remove from the
plot, regardless of p-value filtering.}

\item{progress_bar}{Show a progress bar when testing associations? Useful
for very large datasets.}
}
\value{
The output of plot_grid, a set of plots showing variance explained
and associations between the columns of \code{a} and the principal components
of \code{b}.
}
\description{
Graphical test between principal components and a table of covariates.
}
\examples{
mat <- matrix(rnorm(1000), ncol = 10)
pc <- prcomp(mat)
pca_association_plot(mat, pc)
## only the top 5 associations
pca_association_plot(mat, pc, n = 5)
}
